RHR-23

+++ hr-app/src/main/java/com/remark_herlan/hr_app/dao/RolesDao.java
@@ -12,6 +12,6 @@ import com.remark_herlan.hr_app.model.Roles;
  */
 
 @Repository
-public interface RolesDao extends JpaRepository<Roles, Integer> {
+public interface RolesDao extends JpaRepository<Roles, Long> {
 
 }
